+++ CHANGELOG.md
@@ -2,3 +2,13 @@
 
 * TODO: Describe initial release.
 Initial Release For Testing The Package
+
+## 0.0.2
+
+* TODO: Version 0.0.2.
+Added Default TabBar
+
+## 0.0.3
+
+* TODO: Version 0.0.3.
+Added Custom TabBar